+++ packages/block-editor/src/components/block-list/use-block-props/index.native.js
@@ -1,0 +1,10 @@
+/**
+ * WordPress dependencies
+ */
+import { __unstableGetBlockProps as getBlockProps } from '@wordpress/blocks';
+
+export function useBlockProps( props = {} ) {
+	return props;
+}
+
+useBlockProps.save = getBlockProps;
